Added logging to the install-gcc script.

+++ scripts/remove-installed-gcc-and-dependencies.sh
@@ -5,7 +5,11 @@
 ####################################
 rm -rf /opt/tmp/b-arm-cross/*
 rm -rf /opt/tmp/b-native/*
-rm /opt/tmp/did-reinstall-of-native-binutils
+
+if [ -f /opt/tmp/did-reinstall-of-native-binutils ];
+then
+    rm /opt/tmp/did-reinstall-of-native-binutils
+fi
 
 rm -rf /opt/gcc-4.9-native/*
 rm -rf /opt/gcc-4.9-arm/*
